Add getAmazonAdid() and make symbols visible for all platforms
Previously, some symbols like onResume() were restricted only to Android
platform. This will not allow a cross-platform app since each codebase
will be different or surrounded with if-statements to check for
platforms.
The change now is to allow the symbols for platform-specific functions
to appear in all platforms and limit their functionality based on the
specific platform.

Ex: I can call onResume() in an iOS codebase but it won't have any
effect.